Remove ShowViewModel in favor for MvxNavigationService
